bug related to tests fixxed

+++ src/database/knex/migrations/20240531002027_create-users.ts
@@ -1,15 +1,21 @@
 import type { Knex } from 'knex'
 
 export async function up(knex: Knex): Promise<void> {
-  await knex.schema.createTable('users', (table) => {
-    table.uuid('id').primary().index()
-    table.text('name').notNullable()
-    table.text('email').notNullable()
-    table.text('password').notNullable()
-    table.timestamp('created_at').defaultTo(knex.fn.now()).notNullable()
-  })
+  const exists = await knex.schema.hasTable('users')
+  if (!exists) {
+    await knex.schema.createTable('users', (table) => {
+      table.uuid('id').primary().index()
+      table.text('name').notNullable()
+      table.text('email').notNullable().unique()
+      table.text('password').notNullable()
+      table.timestamp('created_at').defaultTo(knex.fn.now()).notNullable()
+    })
+  }
 }
 
 export async function down(knex: Knex): Promise<void> {
-  await knex.schema.dropTable('users')
+  const exists = await knex.schema.hasTable('users')
+  if (exists) {
+    await knex.schema.dropTable('users')
+  }
 }
