+++ src/database/knex/migrations/20240601223033_create-meals.ts
@@ -1,19 +1,25 @@
 import type { Knex } from 'knex'
 
 export async function up(knex: Knex): Promise<void> {
-  await knex.schema.createTable('meals', (table) => {
-    table.uuid('id').primary()
-    table.uuid('user_id').references('users.id').onDelete('CASCADE').index()
+  const exists = await knex.schema.hasTable('meals')
+  if (!exists) {
+    await knex.schema.createTable('meals', (table) => {
+      table.uuid('id').primary()
+      table.uuid('user_id').references('users.id').onDelete('CASCADE').index()
 
-    table.text('name').notNullable()
-    table.text('description').notNullable()
+      table.text('name').notNullable()
+      table.text('description').notNullable()
 
-    table.text('date').notNullable()
-    table.text('time').notNullable()
-    table.boolean('on_diet').notNullable()
-  })
+      table.text('date').notNullable()
+      table.text('time').notNullable()
+      table.boolean('on_diet').notNullable()
+    })
+  }
 }
 
 export async function down(knex: Knex): Promise<void> {
-  await knex.schema.dropTable('meals')
+  const exists = await knex.schema.hasTable('meals')
+  if (exists) {
+    await knex.schema.dropTable('meals')
+  }
 }
